#19 employee feature done
