--- src/tt.txt
@@ -1,3 +1,0 @@
-안녕하세요
-반갑습니다.
-너무 힘든 월요일이네요
