+++ src/user/repository/UserRepositoryImpl.java
@@ -1,77 +1,162 @@
 package user.repository;
 
-import db.DataBase;
 import db.DB;
-import db.DataBaseImpl;
+import user.domain.Income;
+import user.domain.Jobs;
 import user.domain.User;
 
+import java.sql.*;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.TreeMap;
 
 public class UserRepositoryImpl implements UserRepository {
 
     static class Holder {
         static final UserRepository INSTANCE = new UserRepositoryImpl();
     }
+
     public static UserRepository getInstance() {
         return Holder.INSTANCE;
     }
 
-    DataBase db = DataBaseImpl.getInstance();
-    TreeMap<Integer,User> table = db.getUserTable();
-
     @Override
     public List<User> findAllUser() throws Exception {
-        return table.values().stream().toList();
-    }
-    @Override
-    public User findUserById(int uid) throws Exception {
-        return table.get(uid);
+        List<User> users = new ArrayList<>();
+        String sql = "SELECT * FROM User";
+        try (Connection conn = DB.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql);
+             ResultSet rs = stmt.executeQuery()) {
+            while (rs.next()) {
+                users.add(extractUser(rs));
+            }
+        }
+        return users;
     }
 
     @Override
-    public User findUserByLoginId(String id) throws Exception {
-        for (Map.Entry<Integer, User> entry : table.entrySet()) {
-            User user = entry.getValue();
-            if(user.getLoginId().equals(id)) return user;
+    public User findUserById(int uid) throws Exception {
+        String sql = "SELECT * FROM User WHERE uid = ?";
+        try (Connection conn = DB.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setInt(1, uid);
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (rs.next()) {
+                    return extractUser(rs);
+                }
+            }
         }
         throw new Exception("찾는 아이디가 없습니다.");
     }
 
     @Override
+    public User findUserByLoginId(String loginId) throws Exception {
+        String sql = "SELECT * FROM User WHERE loginId = ?";
+        try (Connection conn = DB.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, loginId);
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (rs.next()) {
+                    return extractUser(rs);
+                }
+            }
+        }
+        throw new Exception("찾는 아이디가 존재하지 않습니다");
+    }
+
+    @Override
     public User createUser(User user) throws Exception {
-        int autoId = table.size();
-        LocalDateTime now = LocalDateTime.now();
-        user.setUid(autoId);
-        user.setCreatedTime(now);
-        user.setDeleteTime(null);
-        table.put(autoId,user);
+        String sql = "INSERT INTO User (loginId, password, name, address, phoneNumber, email, job, income, asset, birth, createdTime) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+        try (Connection conn = DB.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS)) {
+            stmt.setString(1, user.getLoginId());
+            stmt.setString(2, user.getPassword());
+            stmt.setString(3, user.getName());
+            stmt.setString(4, user.getAddress());
+            stmt.setString(5, user.getPhoneNumber());
+            stmt.setString(6, user.getEmail());
+            stmt.setString(7, user.getJob().name());
+            stmt.setString(8, user.getIncome().name());
+            stmt.setInt(9, user.getAsset());
+            stmt.setDate(10, java.sql.Date.valueOf(user.getBirth()));
+            stmt.setTimestamp(11, java.sql.Timestamp.valueOf(user.getCreatedTime()));
+
+            int affectedRows = stmt.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("유저 생성에 실패하였습니다.");
+            }
+
+            try (ResultSet generatedKeys = stmt.getGeneratedKeys()) {
+                if (generatedKeys.next()) {
+                    user.setUid(generatedKeys.getInt(1));
+                } else {
+                    throw new SQLException("유저 생성에 실패하였습니다.");
+                }
+            }
+        }
         return user;
     }
 
     @Override
     public void deleteUser(User user) throws Exception {
-        User target = table.get(user.getUid());
-        if (target != null) {
-            target.setDeleteTime(LocalDateTime.now());
-            table.remove(user.getUid());
-            table.put(user.getUid(), target);
-        } else {
-            throw new Exception("해당 유저가 없습니다.");
+        String sql = "DELETE FROM User WHERE uid = ?";
+        try (Connection conn = DB.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setInt(1, user.getUid());
+            int affectedRows = stmt.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("계정 탈퇴중 에러가 발생했습니다..");
+            }
         }
     }
 
     @Override
-    public void changePassword(User user, String newPasswrod) throws Exception {
-        User target = table.get(user.getUid());
-        if (target != null) {
-            target.setPassword(newPasswrod);
-            table.remove(user.getUid());
-            table.put(user.getUid(), target);
-        } else {
-            throw new Exception("비밀번호 변경에 실패하였습니다.");
+    public void changePassword(User user, String newPassword) throws Exception {
+        String sql = "UPDATE User SET password = ? WHERE uid = ?";
+        try (Connection conn = DB.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, newPassword);
+            stmt.setInt(2, user.getUid());
+            int affectedRows = stmt.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("계정생성에 실패하였습니다.");
+            }
+        }
+    }
+
+    private User extractUser(ResultSet rs) throws Exception {
+        try {
+            User user = new User(
+                    rs.getString("loginId"),
+                    rs.getString("password"),
+                    rs.getString("name"),
+                    rs.getString("address"),
+                    rs.getString("phoneNumber"),
+                    rs.getString("email"),
+                    Jobs.valueOf(rs.getString("job")),
+                    Income.valueOf(rs.getString("income")),
+                    rs.getInt("asset"),
+                    rs.getString("birth")
+            );
+
+            user.setUid(rs.getInt("uid"));
+            Timestamp createdTime = rs.getTimestamp("createdTime");
+            LocalDateTime localDateTime = null;
+            if (createdTime != null) {
+                localDateTime = createdTime.toLocalDateTime();
+                user.setCreatedTime(localDateTime);
+            }
+
+            Timestamp deleteTime = rs.getTimestamp("deleteTime");
+            LocalDateTime createdTimeToLocal = null;
+            if (deleteTime != null) {
+                createdTimeToLocal = deleteTime.toLocalDateTime();
+                user.setDeleteTime(createdTimeToLocal);
+            }
+
+            return user;
+        } catch (Exception e) {
+            throw new Exception(e.getMessage());
         }
     }
 }
